Backtest scalping apertura DAX -> rejected (no edge dopo costi)
Implementata la strategia dell'utente in versione meccanica + simulatore intraday
dedicato (sessione, finestre orarie, un setup/giorno, elasticita' d'ingresso con
fascia di prossimita', parziali + trailing, sizing a rischio %).

- src/backtest/dax_open.py: strategia + backtest a sessione (find_signal, run_day,
  gestione trade, metriche). Parametri elasticita' scelti come compromesso motivato.
- tests/test_dax_open.py: logica verificata su giornata sintetica (segnale, long su
  rimbalzo, giorno piatto). 23/23 verdi.
- fix adapter dukascopy: volume tenuto in FLOAT (sugli indici e' frazionario;
  troncarlo a int azzerava il segnale volume). Alias indici gia' presenti.

Risultati (DAX M1, 3y, Dukascopy):
- in-sample spread 1,5pt: PF 1,045, win 39%, +9,3% in 3y -> edge TENUE;
- sensibilita' spread: PF 0,976 a 2,5pt (costo realistico in apertura) -> muore;
- incostante per anno (2023 in perdita); candele "oro" (vol_mult alto) peggiorano;
- 1 variante d'uscita su 5 resta positiva (rumore, non OOS).
Evidenze in raw/exp_2026-07-06_dax_open_backtest.json (immutabile).

Gate: rejected come strategia meccanica automatica (break-even dopo costi).
L'edge reale dell'utente e' probabilmente discrezionale (timing/selezione), non
catturabile da questa regola fissa. Wiki aggiornata (strategy/instrument/exp/index/log).

+++ src/adapters/dukascopy/data.py
@@ -92,7 +92,9 @@ def _df_to_records(df) -> list[dict]:
                 "high": float(row["high"]),
                 "low": float(row["low"]),
                 "close": float(row["close"]),
-                "volume": int(row["volume"]) if row["volume"] == row["volume"] else 0,
+                # volume TENUTO in virgola mobile: sugli indici Dukascopy è
+                # frazionario (<1). Troncarlo a int azzererebbe il segnale.
+                "volume": float(row["volume"]) if row["volume"] == row["volume"] else 0.0,
             }
         )
     return out
